Exceptions別の構成に変更
