--- src/main/java/com/example/name/exception/CustomExceptions.java
@@ -1,21 +1,0 @@
-package com.example.name.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-public class CustomExceptions {
-
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public static class NotFoundException extends RuntimeException {
-        public NotFoundException(String message) {
-            super(message);
-        }
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public static class BadRequestException extends RuntimeException {
-        public BadRequestException(String message) {
-            super(message);
-        }
-    }
-}
